Apply patch from Jason Smith - RT#46495
Comment from RT:

I think this may be the same problem reported in bug #31291, but I don't
think that is the correct way to fix the problem. Instead of commenting
out the split in the original parsing set attribute function, I made a
patch for the register function that doesn't try to split an array ref
that was already split earlier.