Apply patch from Jason Smith - RT#46497
Comment from RT:

The config parser cannot handle string attributes that can contain an
arbitrary string of characters, including a right curly bracket [}].
The parser incorrectly assumes that this is the end of the block
definition and the remaining attributes for that definition are left
unassigned.

I created a very simple patch for this bug that will probably work in
most cases, but it assumes that the right curly bracket that closes a
definition block contains no more important data or comments after it on
the same line and is only followed by zero or more white spaces. This
was just the simplest solution that worked for our Nagios config, but a
proper fix may require a more complicated parser rewrite to be context
sensitive and allow right curly brackets within string attributes.