Allow for ->name() to not be unique on service objects.
find_object() expects the Object to be unique per the name() method.
This is not true for Service Objects, you can have multiple Service
Objects with the same service_description. The classic example is
having a CPU utilization test for both production and development
servers, but the notification_period may be different for those
different set of hosts.

Thanks to PIRZYK (RT#63806)