notes field does not support having a ';' in it
If the notes field has a semi-colon listed in it, Nagios::Object strips
out the rest of the line (treating it as a comment). This hack is to
put the entire notes field back in.

When viewing the notes field content in the nagios app, the comment
character is displayed on the screen (not stripped out).

Not sure what other fields in nagios itself has this syntax (allowing
the comment character as a valid field content).

Thanks to PIRZYK (RT#66348)